refactored swagger codegen imports.
