+++ src/main/java/io/swagger/codegen/languages/DefaultCodegenConfig.java
@@ -2,19 +2,19 @@ package io.swagger.codegen.languages;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.github.jknack.handlebars.Handlebars;
-import io.swagger.codegen.CliOption;
-import io.swagger.codegen.CodegenArgument;
-import io.swagger.codegen.CodegenConfig;
-import io.swagger.codegen.CodegenConstants;
-import io.swagger.codegen.CodegenModel;
-import io.swagger.codegen.CodegenModelFactory;
-import io.swagger.codegen.CodegenModelType;
-import io.swagger.codegen.CodegenOperation;
-import io.swagger.codegen.CodegenParameter;
-import io.swagger.codegen.CodegenProperty;
-import io.swagger.codegen.CodegenResponse;
-import io.swagger.codegen.CodegenSecurity;
-import io.swagger.codegen.SupportingFile;
+import io.swagger.codegen.v3.CliOption;
+import io.swagger.codegen.v3.CodegenArgument;
+import io.swagger.codegen.v3.CodegenConfig;
+import io.swagger.codegen.v3.CodegenConstants;
+import io.swagger.codegen.v3.CodegenModel;
+import io.swagger.codegen.v3.CodegenModelFactory;
+import io.swagger.codegen.v3.CodegenModelType;
+import io.swagger.codegen.v3.CodegenOperation;
+import io.swagger.codegen.v3.CodegenParameter;
+import io.swagger.codegen.v3.CodegenProperty;
+import io.swagger.codegen.v3.CodegenResponse;
+import io.swagger.codegen.v3.CodegenSecurity;
+import io.swagger.codegen.v3.SupportingFile;
 import io.swagger.codegen.handlebars.helpers.BaseItemsHelper;
 import io.swagger.codegen.handlebars.helpers.BracesHelper;
 import io.swagger.codegen.handlebars.helpers.HasHelper;
@@ -22,7 +22,7 @@ import io.swagger.codegen.handlebars.helpers.HasNotHelper;
 import io.swagger.codegen.handlebars.helpers.IsHelper;
 import io.swagger.codegen.handlebars.helpers.IsNotHelper;
 import io.swagger.codegen.handlebars.helpers.StringUtilHelper;
-import io.swagger.codegen.utils.ModelUtils;
+import io.swagger.codegen.v3.utils.ModelUtils;
 import io.swagger.v3.core.util.Json;
 import io.swagger.v3.core.util.Yaml;
 import io.swagger.v3.oas.models.OpenAPI;
@@ -84,18 +84,18 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
-import static io.swagger.codegen.CodegenConstants.HAS_ONLY_READ_ONLY_EXT_NAME;
-import static io.swagger.codegen.CodegenConstants.HAS_OPTIONAL_EXT_NAME;
-import static io.swagger.codegen.CodegenConstants.HAS_REQUIRED_EXT_NAME;
-import static io.swagger.codegen.CodegenConstants.IS_ARRAY_MODEL_EXT_NAME;
-import static io.swagger.codegen.CodegenConstants.IS_CONTAINER_EXT_NAME;
-import static io.swagger.codegen.CodegenConstants.IS_ENUM_EXT_NAME;
+import static io.swagger.codegen.v3.CodegenConstants.HAS_ONLY_READ_ONLY_EXT_NAME;
+import static io.swagger.codegen.v3.CodegenConstants.HAS_OPTIONAL_EXT_NAME;
+import static io.swagger.codegen.v3.CodegenConstants.HAS_REQUIRED_EXT_NAME;
+import static io.swagger.codegen.v3.CodegenConstants.IS_ARRAY_MODEL_EXT_NAME;
+import static io.swagger.codegen.v3.CodegenConstants.IS_CONTAINER_EXT_NAME;
+import static io.swagger.codegen.v3.CodegenConstants.IS_ENUM_EXT_NAME;
 import static io.swagger.codegen.handlebars.helpers.ExtensionHelper.getBooleanValue;
 import static io.swagger.codegen.languages.CodegenHelper.getDefaultIncludes;
 import static io.swagger.codegen.languages.CodegenHelper.getImportMappings;
 import static io.swagger.codegen.languages.CodegenHelper.getTypeMappings;
 import static io.swagger.codegen.languages.CodegenHelper.initalizeSpecialCharacterMapping;
-import static io.swagger.codegen.utils.ModelUtils.processCodegenModels;
+import static io.swagger.codegen.v3.utils.ModelUtils.processCodegenModels;
 
 public abstract class DefaultCodegenConfig implements CodegenConfig {
     private static final Logger LOGGER = LoggerFactory.getLogger(DefaultCodegenConfig.class);
