+++ src/main/java/io/swagger/codegen/languages/html/StaticDocCodegen.java
@@ -1,10 +1,10 @@
 package io.swagger.codegen.languages.html;
 
-import io.swagger.codegen.CliOption;
-import io.swagger.codegen.CodegenConfig;
-import io.swagger.codegen.CodegenConstants;
-import io.swagger.codegen.CodegenType;
-import io.swagger.codegen.SupportingFile;
+import io.swagger.codegen.v3.CliOption;
+import io.swagger.codegen.v3.CodegenConfig;
+import io.swagger.codegen.v3.CodegenConstants;
+import io.swagger.codegen.v3.CodegenType;
+import io.swagger.codegen.v3.SupportingFile;
 import io.swagger.codegen.languages.DefaultCodegenConfig;
 import org.apache.commons.lang3.StringUtils;
 import java.io.File;
