+++ src/main/java/io/swagger/codegen/languages/java/JavaJAXRSCXFCDIServerCodegen.java
@@ -1,9 +1,9 @@
 package io.swagger.codegen.languages.java;
 
-import io.swagger.codegen.CliOption;
-import io.swagger.codegen.CodegenModel;
-import io.swagger.codegen.CodegenProperty;
-import io.swagger.codegen.SupportingFile;
+import io.swagger.codegen.v3.CliOption;
+import io.swagger.codegen.v3.CodegenModel;
+import io.swagger.codegen.v3.CodegenProperty;
+import io.swagger.codegen.v3.SupportingFile;
 import io.swagger.codegen.languages.features.BeanValidationFeatures;
 import org.apache.commons.lang3.StringUtils;
 import java.io.File;
