+++ src/main/java/io/swagger/codegen/languages/kotlin/KotlinClientCodegen.java
@@ -1,9 +1,9 @@
 package io.swagger.codegen.languages.kotlin;
 
-import io.swagger.codegen.CliOption;
-import io.swagger.codegen.CodegenConstants;
-import io.swagger.codegen.CodegenType;
-import io.swagger.codegen.SupportingFile;
+import io.swagger.codegen.v3.CliOption;
+import io.swagger.codegen.v3.CodegenConstants;
+import io.swagger.codegen.v3.CodegenType;
+import io.swagger.codegen.v3.SupportingFile;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
