+++ src/main/java/io/swagger/codegen/languages/swift/Swift3Codegen.java
@@ -1,7 +1,7 @@
 package io.swagger.codegen.languages.swift;
 
-import io.swagger.codegen.CliOption;
-import io.swagger.codegen.CodegenConstants;
+import io.swagger.codegen.v3.CliOption;
+import io.swagger.codegen.v3.CodegenConstants;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.Arrays;
