+++ src/main/java/io/swagger/codegen/languages/swift/Swift4Codegen.java
@@ -1,10 +1,10 @@
 package io.swagger.codegen.languages.swift;
 
-import io.swagger.codegen.CliOption;
-import io.swagger.codegen.CodegenConstants;
-import io.swagger.codegen.CodegenModel;
-import io.swagger.codegen.CodegenProperty;
-import io.swagger.codegen.SupportingFile;
+import io.swagger.codegen.v3.CliOption;
+import io.swagger.codegen.v3.CodegenConstants;
+import io.swagger.codegen.v3.CodegenModel;
+import io.swagger.codegen.v3.CodegenProperty;
+import io.swagger.codegen.v3.SupportingFile;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.Arrays;
