+++ src/main/java/io/swagger/codegen/languages/typescript/AbstractTypeScriptClientCodegen.java
@@ -1,12 +1,12 @@
 package io.swagger.codegen.languages.typescript;
 
-import io.swagger.codegen.CliOption;
-import io.swagger.codegen.CodegenConfig;
-import io.swagger.codegen.CodegenConstants;
-import io.swagger.codegen.CodegenModel;
-import io.swagger.codegen.CodegenProperty;
-import io.swagger.codegen.CodegenType;
-import io.swagger.codegen.DefaultCodegen;
+import io.swagger.codegen.v3.CliOption;
+import io.swagger.codegen.v3.CodegenConfig;
+import io.swagger.codegen.v3.CodegenConstants;
+import io.swagger.codegen.v3.CodegenModel;
+import io.swagger.codegen.v3.CodegenProperty;
+import io.swagger.codegen.v3.CodegenType;
+import io.swagger.codegen.v3.DefaultCodegen;
 import io.swagger.codegen.languages.DefaultCodegenConfig;
 import io.swagger.codegen.languages.java.JavaClientCodegen;
 import io.swagger.v3.oas.models.media.ArraySchema;
@@ -30,7 +30,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.TreeSet;
 
-import static io.swagger.codegen.CodegenConstants.IS_ENUM_EXT_NAME;
+import static io.swagger.codegen.v3.CodegenConstants.IS_ENUM_EXT_NAME;
 import static io.swagger.codegen.languages.helpers.ExtensionHelper.getBooleanValue;
 
 public abstract class AbstractTypeScriptClientCodegen extends DefaultCodegenConfig {
