+++ src/main/java/io/swagger/codegen/languages/typescript/TypeScriptAngularClientCodegen.java
@@ -10,12 +10,12 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
 
-import io.swagger.codegen.CliOption;
-import io.swagger.codegen.CodegenModel;
-import io.swagger.codegen.CodegenParameter;
-import io.swagger.codegen.CodegenOperation;
-import io.swagger.codegen.SupportingFile;
-import io.swagger.codegen.utils.SemVer;
+import io.swagger.codegen.v3.CliOption;
+import io.swagger.codegen.v3.CodegenModel;
+import io.swagger.codegen.v3.CodegenParameter;
+import io.swagger.codegen.v3.CodegenOperation;
+import io.swagger.codegen.v3.SupportingFile;
+import io.swagger.codegen.v3.utils.SemVer;
 import io.swagger.v3.oas.models.media.ArraySchema;
 import io.swagger.v3.oas.models.media.FileSchema;
 import io.swagger.v3.oas.models.media.MapSchema;
